Projekt Ochrona Danych/flask/requirements.txt
Flask
gunicorn
uwsgi
passlib
flask_wtf
pycryptodome
pycryptodomex
flask-sqlalchemy